refactor(app): RoundTab is a shared component
